--- src/main/java/chess/domain/board/BoardInfo.java
@@ -1,6 +1,0 @@
-package chess.domain.board;
-
-public class BoardInfo {
-    public static final int BOARD_MAX_INDEX = 7;
-    public static final int BOARD_MIN_INDEX = 0;
-}
