+++ src/test/java/chess/domain/board/BoardFactoryTest.java
@@ -1,7 +1,5 @@
-package chess.factory;
+package chess.domain.board;
 
-import chess.domain.board.Board;
-import chess.domain.board.BoardFactory;
 import chess.domain.game.GameStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
